Moving function to their respective classes and looping the main using timespan objects.
