+++ MuseumTours/Customer.cs
@@ -1,3 +1,4 @@
+using Newtonsoft.Json;
 public class Customer
 {
     public string CustomerCode;
@@ -5,5 +6,20 @@ public class Customer
     public Customer(string customercode)
     {
         CustomerCode = customercode;
+    }
+    public bool CheckIfCustomerInList(string idcustomer)
+    {
+        using StreamReader reader = new("../../../Customers.Json");
+        string File2Json = reader.ReadToEnd();
+        List<Customer> listOfCustomers = JsonConvert.DeserializeObject<List<Customer>>(File2Json)!;
+
+        foreach (Customer customer in listOfCustomers)
+        {
+        if (customer.CustomerCode == idcustomer)
+        {
+            return true;
+        }
+        }
+        return false;
     }
 }
