+++ MuseumTours/Tours.cs
@@ -20,17 +20,54 @@ public class Tours
         Time = time;
     }
 
-    public void CheckIn()
+    public static bool AddToTour(string customerid, string tourid)
     {
-        Console.WriteLine("U heeft succesvol gereserveerd.");
-    }
+    string filePath = "../../../Tourslist.Json";
 
-    public void LoadJson()
+    string File2Json = File.ReadAllText(filePath);
+    List<Tours> listOfTours = JsonConvert.DeserializeObject<List<Tours>>(File2Json)!;
+
+    foreach (Tours tour in listOfTours)
     {
-        using (StreamReader r = new StreamReader("Rondleidingen.json"))
-        {
-            string json = r.ReadToEnd();
-            List<Tours>? items = JsonConvert.DeserializeObject<List<Tours>>(json);
-        }
+      if (tour.ID == tourid && tour.Time > DateTime.Now)
+      {
+        Customer customer = new Customer(customerid);
+        tour.Customer_Codes.Add(customer);
+        tour.Spots--;
+        Console.WriteLine($"Reservering geplaatst. U word op {tour.Time} verwacht bij het verzamelpunt.");
+
+        string updatedJson = JsonConvert.SerializeObject(listOfTours, Formatting.Indented);
+        File.WriteAllText(filePath, updatedJson);
+
+        return true;
+      }
     }
+    Console.WriteLine($"We hebben geen tour kunnen vinden met het ingevoerde nummer:{tourid}.");
+    return false;
+    }
+    public static void ShowAvailableTours(string filename)
+    {
+    using StreamReader reader = new(filename);
+    string File2Json = reader.ReadToEnd();
+    List<Tours> listOfTours = JsonConvert.DeserializeObject<List<Tours>>(File2Json)!;
+
+    int Count = 0;
+    Console.WriteLine("Dit zijn de nog beschikbare rondleidingen voor vandaag:");
+    bool touratleast = false;
+    foreach (Tours tour in listOfTours)
+    {
+      if (tour.Time > DateTime.Now && tour.Spots > 0)
+      {
+        Count++;
+        string timeString = tour.Time.ToString("HH:mm");
+        Console.WriteLine($"{tour.ID}. starttijd: {timeString} beschikbare plekken: {tour.Spots}");
+        touratleast = true;
+      }
+    }
+    if (touratleast == false)
+    {
+      Console.WriteLine("Er zijn op het moment geen beschikbare rondleidingen.");
+    }
+  }
 }
+
